backend/requirements.txt
asgiref==3.5.2
channels==4.0.0
Django==4.1.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-dotenv==1.0.0
redis==4.5.5
daphne
django-cors-headers
psycopg2
channels_redis